Refactor grid header into its own component
